requirements.txt
pypng==0.0.14
twitter==1.9.1
